@@ -1,23 +1,10 @@
 const nexumeConfig = {
-	/**
-	 * Metadata
-	 * 
-	 * title:       string
-	 * description: string
-	 * icon:        string,
-	 */
 	title: "John Doe's resume",
-	description: "John Doe's resume using next.js",
-	icon: 'undefined',
-
-	/**
-	 * Layout
-	 * 
-	 * width: 	   number
-	 * themeColor: string
-	 */
+	description: "John Doe's resume using nexume.",
+	icon: '/favicon.ico',
+	profileImg: '/profile_img.png',
 	width: 794,
 	themeColor: '#8A5CF5',
 }
 
-module.exports = nexumeConfig
+export default nexumeConfig
