@@ -1,45 +1,27 @@
-const nexume = require('./nexume.config');
+import nexumeConfig from "./nexume.config"
 
 const config = {
   content: [
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/modules/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/app/**/*.{js,ts,jsx,tsx,mdx}'
   ],
   theme: {
     extend: {
       colors: {
         h: {
-          transparent: 'transparent',
-          main: nexume.themeColor,
+          main: nexumeConfig.themeColor,
         },
         l: {
-          100: '#FFFFFF',
-          200: '#FAFAFA',
-          300: '#E0E0E0',
-          400: '#ABABAB',
-          500: '#848484',
-          600: '#5C5C5C',
-          700: '#222222'
-        },
-        d: {
-          100: '#FFFFFF',
-          200: '#DADADA',
-          300: '#B3B3B3',
-          400: '#666666',
-          500: '#363636',
-          600: '#242424'
+          'e': '#E0E0E0',
+          'c': '#CCCCCC',
+          'a': '#ABABAB',
+          '8': '#848484',
+          '5': '#5C5C5C',
+          '2': '#222222'
         }
-      },
-    },
+      }
+    }
   },
-  plugins: [],
-  safelist: [
-      { pattern: /(px|py|pt|pb|mt|mb|ml|mr)-(0.5|1|1.5|2|2.5|3|3.5|4|5|6|7|8|9|10|11|12|14|16|20|24|28|32|36|40|44|48|52|56|60|64|72|80|96|)/ },
-      { pattern: /(gap)-(0.5|1|1.5|2|2.5|3|3.5|4|5|6|7|8|9|10|11|12|14|16|20|24|28|32|36|40|44|48|52|56|60|64|72|80|96|)/ },
-      { pattern: /(text)-(xs|sm|base|lg|xl|2xl|3xl|4xl|5xl|6xl|7xl|8xl|9xl)/ },
-      { pattern: /(font)-(thin|extralight|light|normal|medium|semibold|bold|extrabold|black)/ },
-      { pattern: /(tracking)-(tighter|tight|normal|wide|wider|widest)/ },
-      { pattern: /(rounded)-(none|sm|md|lg|full)/ },
-  ]
+  plugins: []
 }
+
 export default config
